Add files via upload
Added support for an n log n + n runtime of document recommendation, based on the similarity of any other document in the corpus, for all documents in the corpus.

Brute force comparison of a million documents approximately needs one trillion calculations - 999,999 documents exist to compare to for each document in the collection. We can reduce this one trillion down to approximately 7 million operations.

In this script we perform the following:
Download data from the cloud using metabeaver helper functions. You will need to create your own yaml file or replace the references to project id, dataset and tableset with your own details (or modify the script to work with your own dataframe from another source, e.g. a spreadsheet ingested into pandas).
We create a summary of the text column using a lightweight transformer. 'Page Text' is the default target column - change this to your text column of interest.
After data download, we target the summary column to create a 384 dimensional vector, representing semantic features in the summary. 
We ingest all of the vectors into a data structure where the more similar vectors are organised by their similarity. This could be thought of as an n-dimensional representation of a sort such as "ascending numbers". The data storage structure for our vectors, grouping them by similarity, is sk-learn's BallTree, a variant of a Binary Tree data structure: https://scikit-learn.org/stable/modules/generated/sklearn.neighbors.BallTree.html .
We use the efficient sorted BallTree to search (the n log) part of our algorithm for our n documents. 

Total run time should be approximately n log n + n, or, more formally, approximately nlogn (as we approach infinity.... Weeeeeeeee!).